+++ HISTORY.rst
@@ -3,6 +3,13 @@ HISTORY
 =======
 
 =====================
+1.1.0 (18 February, 2016)
+=====================
+* added the option to run FishTaco with no functional profile input, predicting it from the taxonomic profiles and genomic content inputs
+* added a unit test for the option to run with no functional profiles input
+* added the option to run FishTaco on only a subset of functions by supplying a list file as input
+
+=====================
 1.0.5 (15 July, 2015)
 =====================
 * Initial release